Rename ledmatrix.c to match .c file
